move heap to the process dictionary.
passing the heap as a variable in functions causes an
un-yet-confirmed performance issue. This is likely due to variable
capture in heavily nested lambdas. Moving the heap to the
process dictionary resolves these issues. Use a wrapper function
in case the heap will move to another part of the BEAM.