delay creating root env.
Mark and sweep needs a root environment to start.
When creating a heap and root environment at first instatiation of
the HOF rtned by load_block, we end up w/ a heap with two empty
envs at top of tree, ie
    root (empty)
     |
    child (empty)
     |
    grandchild (actual root env)
because envs track their parents, but not their children, finding
ancestors is easy, but finding decendants is hard.
By delaying instatiation of the heap, and the root env until the
2nd child block is encountered properly sets the root env for the
runtime case.
not yet sure if this will break the compiler/other bytecode later.