program rewrite, testing focused, see below.
Getting GC working requires a significant advancement in my personal
understanding of how the virtual machine functions. Experimental
versions of the library took place in external repos. These are being
re-integrated into the primary repo (and will then be deleted).

This seeks to
    1.) take a "testing first" philosophy
    2.) reduce code complexity and size
    3.) prepare the core library architecture for proper GC

The core primitives of BQN will eventually be remerged into this branch.